requirements.txt
asgiref==3.3.1
Django==3.0.3
django-extensions==3.1.1
django-ninja==0.11.0
djangorestframework==3.12.2
pydantic==1.7.3
PyJWT==2.0.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
